+++ src/Api/Analysers/MinJumpAchievement.cs
@@ -5,7 +5,7 @@ namespace Api.Analysers {
     public class MinJumpAchievement {
         public int JumpAchievementId { get; set; }
         public string Name { get; set; }
-        public decimal MinAirtime { get; set; }
+        public double MinAirtime { get; set; }
 
         internal bool Check(IEnumerable<RideJumpAnalysis> jumps) {
             return jumps.Any(i => i.Airtime >= MinAirtime);
