+++ src/DataAccess/Models/RideLocation.cs
@@ -13,11 +13,11 @@ namespace DataAccess.Models
         public int RideLocationId { get; set; }
         public int RideId { get; set; }
         public DateTime Timestamp { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
-        public decimal AccuracyInMetres { get; set; }
-        public decimal SpeedMetresPerSecond { get; set; }
-        public decimal Altitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double AccuracyInMetres { get; set; }
+        public double SpeedMetresPerSecond { get; set; }
+        public double Altitude { get; set; }
 
         public virtual Ride Ride { get; set; }
         public virtual ICollection<SegmentAttemptLocation> SegmentAttemptLocation { get; set; }
