+++ src/Database/Create/Tables/AccelerometerReading.sql
@@ -4,7 +4,7 @@
     [RideId] int NOT NULL CONSTRAINT [FK_AccelerometerReading_Ride] REFERENCES [Ride],
 
     [Time] datetime NOT NULL,
-    [X] decimal(5,3) NOT NULL,
-    [Y] decimal(5,3) NOT NULL,
-    [Z] decimal(5,3) NOT NULL,
+    [X] float NOT NULL,
+    [Y] float NOT NULL,
+    [Z] float NOT NULL,
 )
