+++ src/Database/Create/Tables/JumpAchievement.sql
@@ -2,5 +2,5 @@
     [JumpAchievementId] int NOT NULL IDENTITY(1,1) PRIMARY KEY,
 
     [Name] nvarchar(200) NOT NULL,
-    [MinAirtime] decimal(5,3) NOT NULL,
+    [MinAirtime] float NOT NULL,
 )
