+++ src/Database/Create/Tables/SpeedAchievement.sql
@@ -2,5 +2,5 @@
     [SpeedAchievementId] int NOT NULL IDENTITY(1,1) PRIMARY KEY,
 
     [Name] nvarchar(200) NOT NULL,
-    [MinMph] decimal(4,1) NOT NULL,
+    [MinMph] float NOT NULL,
 )
